refactored the architecture to utilize single websocket for all operations in the chat
