+++ live-chat/package.json
@@ -8,6 +8,8 @@
   "main": "dist/index.js",
   "scripts": {
     "dev": "tsx watch src/main.ts",
+    "build": "tsc",
+    "start": "node dist/main.js",
     "test": "vitest"
   },
   "dependencies": {
