+++ live-chat/vitest.config.ts
@@ -2,6 +2,7 @@ import { defineConfig } from 'vitest/config';
 
 export default defineConfig({
   test: {
-    environment: 'node'
+    environment: 'node',
+    include: ['tests/**/*.test.ts']
   }
 });
